Reorganized folders in Traction project, minor diagnostic refactoring
