--- Traction/Contracts/Providers/MasterContractProvider.cs
@@ -1,18 +1,0 @@
-﻿using System.Collections.Generic;
-using System.Linq;
-
-namespace Traction.Contracts {
-
-    public class MasterContractProvider : IContractProvider {
-
-        static MasterContractProvider() {
-            Instance = new MasterContractProvider();
-        }
-
-        public static MasterContractProvider Instance { get; }
-
-        public IEnumerable<Contract> Contracts =>
-            DefaultValueContractProvider.Instance.Contracts.Concat(
-            BasicComparisonContractProvider.Instance.Contracts);
-    }
-}
