+++ Traction/RoslynExtensions/Semantics/AttributeDataExtensions.cs
@@ -1,7 +1,6 @@
 ﻿using Microsoft.CodeAnalysis;
-using Traction.Roslyn;
 
-namespace Traction.Contracts {
+namespace Traction.Roslyn {
 
     internal static class AttributeDataExtensions {
