+++ Traction/RoslynExtensions/Semantics/IParameterSymbolExtensions.cs
@@ -2,7 +2,7 @@
 using Microsoft.CodeAnalysis;
 using Traction.Roslyn;
 
-namespace Traction.Contracts {
+namespace Traction.Roslyn {
 
     static class IParameterSymbolExtensions {
