+++ Traction/RoslynExtensions/Semantics/SymbolExtensionMethodsImpl.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using Microsoft.CodeAnalysis;
 using Traction.Roslyn;
 
-namespace Traction.Contracts {
+namespace Traction.Roslyn {
 
     internal static class SymbolExtensionMethodsImpl {
